Move TypeReference to common package and update imports in ComponentDefinition and tests
