+++ messaging/src/main/java/org/axonframework/configuration/ComponentDefinition.java
@@ -18,6 +18,7 @@ package org.axonframework.configuration;
 
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
+import org.axonframework.common.TypeReference;
 
 import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
